# backend/internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Touchline API
  version: 0.1.0
  description: |
    Touchline is a text-first football management game. The backend is a single
    binary (`cmd/touchline serve` runs the API, the scheduler world clock, and
    the event worker in one process; the api/scheduler/worker subcommands split
    them for prod isolation).

    Authentication is cookie-based: login exchanges credentials for a rotating
    access/refresh cookie pair, and every game endpoint below a security scheme
    marker requires a valid `access_token` cookie. The whole world is scoped to
    the session, so "my world" reads are always the caller's own world.

    Launch model: worlds are created as metadata (name), then an admin declares
    countries and leagues, and finally seeds the whole world with clubs,
    players, and memberships. Seasons/league play follow the fixture system
    driven by the scheduler.

    The interactive reference is this page; the raw spec lives at
    `/openapi.yaml`. Live game state (world ticks, match pace 15-min ticks)
    streams over `/ws` (see the WebSocket section).
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (cmd/touchline serve)
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the session
      operationId: refresh
      security: []
      description: |
        Exchanges the `refresh_token` cookie for a fresh cookie pair for the same
        manager. Used when the access token expires.
      responses:
        "200":
          description: Session rotated; new cookie pair set.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
              example: { status: ok }
        "401":
          description: Missing/invalid refresh token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "missing refresh token" }
...
  /api/clubs/{id}/ledger:
    get:
      tags: [finance]
      summary: Club ledger entries
      operationId: getLedger
      description: Append-only financial entries (genesis, weekly wages, transfers). Most-recent first.
      parameters:
        - $ref: "#/components/parameters/ClubID"
      responses:
        "200":
          description: Ledger entry list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LedgerEntry" }
              examples:
                sample:
                  value:
                    - id: 00000000-0000-4000-8000-000000000010
                      entry_type: GENESIS
                      category: startup_capital
                      amount: 35000000
                      description: Starting cash at club creation
                      occurred_at: "2026-06-01T19:00:00Z"
                    - id: 00000000-0000-4000-8000-000000000011
                      entry_type: WAGE
                      category: wage
                      amount: -1264000
                      description: Weekly wages for 24 players
                      related_event_id: 00000000-0000-4000-8000-000000000012
                      occurred_at: "2026-06-14T19:00:00Z"
        "403":
          description: Not the owning manager.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/clubs/{id}/contracts:
    get:
      tags: [finance]
      summary: Club player contracts
      operationId: getContracts
      parameters:
        - $ref: "#/components/parameters/ClubID"
      responses:
        "200":
          description: Active player contracts.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Contract" }
              examples:
                sample:
                  value:
                    - id: 00000000-0000-4000-8000-000000000013
                      player_id: 00000000-0000-4000-8000-00000000000c
                      player_name: Leo Marsh
                      weekly_wage: 45000
                      signing_bonus: 500000
                      start_date: "2026-06-01"
                      end_date: "2028-06-30"
                      release_clause: 25000000
                      status: active
...
  /api/admin/worlds:
    get:
      tags: [admin]
      summary: List worlds
      operationId: listWorlds
      description: Lists every world row (lifecycle metadata only).
      responses:
        "200":
          description: The worlds.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/World" }
    post:
      tags: [admin]
      summary: Create a world
      operationId: createWorld
      description: |
        Registers a new world in a provisioning state. A world is then built by
        declaring countries/leagues and seeding (no playable season exists
        until the start-season service runs).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "Test League World" }
      responses:
        "201":
          description: The created world row.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/World" }
              examples:
                sample:
                  value:
                    id: 00000000-0000-4000-8000-000000000001
                    name: Test League World
                    status: provisioning
                    created_at: "2026-06-01T19:00:00Z"
        "409":
          description: A world with this name already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie set by POST /api/auth/login and rotated by /api/auth/refresh.
  parameters:
...
    ClubID:
      name: id
      in: path
      required: true
      description: Club id.
      schema: { type: string, format: uuid, example: "00000000-0000-4000-8000-000000000005" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: internal error }
...
    Status:
      type: object
      properties:
        status: { type: string, example: ok }
...
    World:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [provisioning, seeding, active, open_beta, paused, archived] }
        created_at: { type: string, format: date-time }
...
    LedgerEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        entry_type: { type: string, enum: [GENESIS, WAGE, TRANSFER_IN, TRANSFER_OUT, PRIZE, SPONSOR] }
        category: { type: string }
        amount: { type: integer, format: int64 }
        description: { type: string }
        related_event_id: { type: string, format: uuid, nullable: true }
        occurred_at: { type: string, format: date-time }
    Contract:
      type: object
      properties:
        id: { type: string, format: uuid }
        player_id: { type: string, format: uuid }
        player_name: { type: string }
        weekly_wage: { type: integer, format: int64 }
        signing_bonus: { type: integer, format: int64 }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        release_clause: { type: integer, format: int64, nullable: true }
        status: { type: string }
...
security:
  - sessionCookie: []